REF: test files for sort_index, sort_values
